moon- 친구페이지 작업

+++ somoim/src/main/java/com/sample/dao/FollowerDao.java
@@ -13,7 +13,7 @@ public interface FollowerDao {
 	List<MoimFollowDto> getFollowersByUserId(String userId);
 	List<MoimJoinUserMoimDto> getUserMoims(String userId);
 	// 팔로우돼있는지 찾기 1 == 팔로우 0 == 안돼있음
-	long findFollower(MoimFollow moimFollow);
+	int findFollower(MoimFollow moimFollow);
 	
 	void deleteFollower(MoimFollow moimFollow);
 }
