+++ somoim/src/main/java/com/sample/service/MypageService.java
@@ -29,7 +29,7 @@ public interface MypageService {
 	// 쪽지 인서트
 	void AddMessage(MoimAlram moimAlram);
 	
-	long followYn(MoimFollow moimFollow);
+	int followYn(MoimFollow moimFollow);
 	
 	
 	void deleteFollower(MoimFollow follow);
